Module argonautica::config

Enums and defaults for Argon2 configuration options (e.g. hash_len, Variant, Version, etc.)

Structs

HasherConfig

Read-only configuration for Hasher. Can be obtained by calling the config method on an instance of Hasher

VerifierConfig

Read-only configuration for Verifier. Can be obtained by calling the config method on an instance of Verifier

Enums

Backend

Enum representing the choice between a C implementation of the Argon2 algorithm or a Rust implementation. Currently only a C implemenation is supported. You will get an error if you choose the Rust backend. The intention is to write a Rust implementation in the future.

Variant

Enum representing the various variants of the Argon2 algorithm ( Argon2d, Argon2i, and Argon2id).

Version

Enum representing the various versions of the Argon2 algorithm ( 0x10 and 0x13)

Constants

DEFAULT_BACKEND

Backend::C

DEFAULT_HASH_LEN

32_u32

DEFAULT_ITERATIONS

192_u32

DEFAULT_MEMORY_SIZE

4096_u32

DEFAULT_OPT_OUT_OF_SECRET_KEY

false

DEFAULT_PASSWORD_CLEARING

false

DEFAULT_SALT_LEN

32_u32

DEFAULT_SECRET_KEY_CLEARING

false

DEFAULT_VARIANT

Variant::Argon2id

DEFAULT_VERSION

Version::_0x13

Functions

default_cpu_pool

Returns a CpuPool with threads equal to the number of logical cores on your machine

default_lanes

Returns the number of logical cores on your machine

default_threads

Returns the number of logical cores on your machine