Enum argonautica::ErrorKind

pub enum ErrorKind {

    AdditionalDataTooLongError,
    BackendEncodeError,
    BackendUnsupportedError,
    Base64DecodeError,
    Bug,
    HashDecodeError,
    HashLenTooShortError,
    HashMissingError,
    IterationsTooFewError,
    LanesTooFewError,
    LanesTooManyError,
    MemoryAllocationError,
    MemorySizeInvalidError,
    MemorySizeTooSmallError,
    OsRngError,
    PasswordImmutableError,
    PasswordMissingError,
    PasswordTooLongError,
    PasswordTooShortError,
    SaltTooLongError,
    SaltTooShortError,
    SecretKeyImmutableError,
    SecretKeyMissingError,
    SecretKeyTooLongError,
    ThreadError,
    ThreadsTooFewError,
    ThreadsTooManyError,
    Utf8EncodeError,
    VariantEncodeError,
    VersionEncodeError,
    // some variants omitted

}

Enum representing the various kinds of errors

Variants

AdditionalDataTooLongError

Additional data too long. Length in bytes must be less than 2^32

BackendEncodeError

Backend encode error. u32 provided could not be encoded into a Backend

BackendUnsupportedError

Rust backend not yet supported. Please use the C backend

Base64DecodeError

Base64 decode error. Bytes provided were invalid base64

Bug

This is a bug in the argonautica crate and should be unreachable. Please file an issue

HashDecodeError

Hash decode error. Hash provided was invalid

HashLenTooShortError

Hash length too short. Hash length must be at least 4

HashMissingError

Hash missing. Attempted to verify without first having provided a hash

IterationsTooFewError

Iterations too few. Iterations must be greater than 0

LanesTooFewError

Lanes too few. Lanes must be greater than 0

LanesTooManyError

Lanes too many. Lanes must be less than 2^24

MemoryAllocationError

C code attempted to allocate memory (using malloc) and failed

MemorySizeInvalidError

Memory size invalid. Memory size must be a power of two

MemorySizeTooSmallError

Memory size too small. Memory size must be at least 8 times the number of lanes

OsRngError

Failed to access OS random number generator

PasswordImmutableError

Password immutable error. You attempted to hash or verify with an immutable password and password_clearing set to true, which is not possible because with an immutable password argonautica cannot zero out the password bytes. To prevent this error, either pass Hasher or Verifier a mutable password or set password_clearing to false

PasswordMissingError

Password missing. Attempted to verify without first having provided a password

PasswordTooLongError

Password too long. Length in bytes must be less than 2^32

PasswordTooShortError

Password too short. Length in bytes must be greater than 0

SaltTooLongError

Salt too long. Length in bytes must be less than 2^32

SaltTooShortError

Salt too short. Length in bytes must be at least 8

SecretKeyImmutableError

Secret key immutable error. You attempted to hash or verify with an immutable secret key and secret_key_clearing set to true, which is not possible because with an immutable secret key argonautica cannot zero out the secret key bytes. To prevent this error, either pass Hasher or Verifier a mutable secret key or set secret_key_clearing to false

SecretKeyMissingError

Secret key missing. Attempted to hash without a secret key without having first opted out of using a secret key

SecretKeyTooLongError

Secret key too long. Length in bytes must be less than 2^32

ThreadError

C code reported a “Threading failure” error

ThreadsTooFewError

Threads too few. Threads must be greater than 0

ThreadsTooManyError

Threads too many. Threads must be less than 2^24

Utf8EncodeError

Utf-8 encode error. Bytes provided could not be encoded into utf-8

VariantEncodeError

Variant encode error. &str provided could not be encoded into a Variant

VersionEncodeError

Version encode error. &str or u32 provided could not be encoded into a Version